V1.0.1 - Completed task content based on the state of lab 1
