+++ src/main/java/com/capgemini/wsb/fitnesstracker/livecoding/bean/MyBean.java
@@ -1,26 +1,25 @@
-package com.capgemini.wsb.fitnesstracker.user.internal;
+package com.capgemini.wsb.fitnesstracker.livecoding.bean;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
-@Service
 @Profile("BeanCycle")
+@Service
 public class MyBean {
 
-    public MyBean () {
+    public MyBean() {
         System.out.println("Instantiation");
     }
 
     @PostConstruct
-    public void init(){
+    public void init() {
         System.out.println("Initializing..");
     }
 
     @PreDestroy
-    public void destroy(){
+    public void destroy() {
         System.out.println("Destroying..");
     }
 }
-
