+++ src/main/java/com/capgemini/wsb/fitnesstracker/user/internal/UserController.java
@@ -26,7 +26,7 @@ class UserController {
     }
 
     @PostMapping
-    public User addUser(@RequestBody UserDto userDto) {
+    public User addUser(@RequestBody UserDto userDto) throws InterruptedException {
 
         // Demonstracja how to use @RequestBody
         System.out.println("User with e-mail: " + userDto.email() + "passed to the request");
