+++ src/main/resources/LAB01.md
@@ -4,35 +4,35 @@ Zadania proszę realizować zgodnie z własnym tempem. Zadania mają pomóc w pr
 
 ## ZADANIE 0
 - Konfiguracja środowiska, wdrożenie do projektu.
-- Konfiguracja prywatnego repozytorium (udostępnienie dostępu prowadzącemu do prywatnego repozytorium z projektem (link) i bieżącym postępem). 
+- Konfiguracja prywatnego repozytorium (udostępnienie dostępu prowadzącemu do 
+- prywatnego repozytorium z projektem (link) i bieżącym postępem). 
 - Hosting: Rekomendacja Github.com
 
 ## ZADANIE 1. Sieciowe API do operacji typu CRUD na klientach
 
 ### Potrzeba biznesowa
 
-Jako użytkownik systemu, chcę mieć możliwość zarządzania użytkownikami serwisu FitnessTracker:
+Jako użytkownik systemu, chcę mieć możliwość zarządzania użytkownikami 
+serwisu FitnessTracker:
 
 - móc ich wyszukiwać, pobierać
 - móc wprowadzać nowych użytkowników do systemu
 - móc usuwać użytkowników z systemu
 - móc aktualizować użytkowników
--
 
 ### Wymagania funkcjonalne
 
 Stworzone API powinno pozwalać na:
 
 - [ ] wylistowanie podstawowych informacji o wszystkich użytkownikach zapisanych w systemie (tylko ID oraz nazwa
-  klienta)
-- [ ] pobranie szczegółów dotyczących wybranego użytkownika (ID, imię, nazwisko, datę urodzenia, e-mail)
+  uzytkownika)
+- [ ] pobranie szczegółów dotyczących wybranego użytkownika (dowolny parametr: ID/imię & nazwisko/datę urodzenia/ e-mail)
 - [ ] utworzenie nowego użytkownika
-- [ ] usunięcie użytkownika
+- [ ] usunięcie użytkownika (konkretny, np. konkretny ID danego uzytkownika)
 - [ ] wyszukiwanie użytkowników po e-mailu, bez rozróżniania wielkości liter, wyszukujące po fragmencie nazwy (zwracane
-  tylko
-  ID oraz e-mail użytkowników)
+  tylko ID oraz e-mail użytkowników)
 - [ ] wyszukiwanie użytkowników po wieku starszym niż zdefiniowany
-- [] aktualizowanie użytkowników w systemie (dowolnie wybrany atrybut)
+- [ ] aktualizowanie użytkowników w systemie (dowolnie wybrany atrybut)
 
 ### Wymagania techniczne
 
@@ -41,10 +41,7 @@ Stworzone API powinno pozwalać na:
   domyślne, pobierające dane za pomocą `findAll()` oraz przetwarzające je za pomocą strumieni (`Stream`). Przykład
   znaleźć można w `UserRepository`
 - [ ] rozwiązanie powinno spełniać zasady SOLID
-- [ ] rozwiązanie powinno być pokryte testami jednostkowymi (>80%)
-- [ ] OPCJONALNE rozwiązanie powinno implementować logikę potrzebną do spełnienia już
-  istniejących [testów integracyjnych API]
-  . NIE należy zmieniać logiki tych testów.
+- [ ] (OPCJONALNIE) rozwiązanie powinno być pokryte testami jednostkowymi (>80%)
 - [ ] rozwiązanie powinno być odpowiednio zhermetyzowane (nie udostępniać funkcjonalności pozostałym pakietom programu)
 - [ ] kod powinien być odpowiednio udokumentowany za pomocą JavaDoc
 - [ ] do kodu powinna zostać dołączona wyeksportowana kolekcja zapytań z programu Postman, pozwalająca przetestować
