+++ src/main/resources/LAB02.md
@@ -31,9 +31,6 @@ Stworzone API powinno pozwalać na:
   znaleźć można w `UserRepository`
 - [ ] rozwiązanie powinno spełniać zasady SOLID
 - [ ] rozwiązanie powinno być pokryte testami jednostkowymi (>80%)
-- [ ] OPCJONALNE rozwiązanie powinno implementować logikę potrzebną do spełnienia już
-  istniejących [testów integracyjnych API]
-  . NIE należy zmieniać logiki tych testów.
 - [ ] rozwiązanie powinno być odpowiednio zhermetyzowane (nie udostępniać funkcjonalności pozostałym pakietom programu)
 - [ ] kod powinien być odpowiednio udokumentowany za pomocą JavaDoc
 - [ ] do kodu powinna zostać dołączona wyeksportowana kolekcja zapytań z programu Postman, pozwalająca przetestować
@@ -67,35 +64,9 @@ Stworzone API powinno pozwalać na:
   znaleźć można w `UserRepository`
 - [ ] rozwiązanie powinno spełniać zasady SOLID
 - [ ] rozwiązanie powinno być pokryte testami jednostkowymi (>80%)
-- [ ] OPCJONALNE rozwiązanie powinno implementować logikę potrzebną do spełnienia już
-  istniejących [testów integracyjnych API]
-  . NIE należy zmieniać logiki tych testów.
 - [ ] rozwiązanie powinno być odpowiednio zhermetyzowane (nie udostępniać funkcjonalności pozostałym pakietom programu)
 - [ ] kod powinien być odpowiednio udokumentowany za pomocą JavaDoc
 - [ ] do kodu powinna zostać dołączona wyeksportowana kolekcja zapytań z programu Postman, pozwalająca przetestować
   stworzone API
 - [ ] rozwiązanie powinno wykorzystywać zwykłe klasy Javowe do definicji obiektów transferu danych (DTO)
 
-## ZADANIE 3. Programowanie aspektowe.
-
-### Potrzeba biznesowa
-
-Jako pracownik utrzymania, chcę mieć możliwość kontroli wykonania się kodu programu za pomocą logów aplikacji.
-
-### Wymagania funkcjonalne
-
-- [ ] być uruchomiony podczas wywoływania metod publicznych serwisów (klas adnotowanych `@Service`)
-- [ ] przed wywołaniem metody logować o niej informację w
-  formacie (`typ zwracany nazwaKlasy.nazwaMetody(typParametru1 nazwaParametru1, ...)`),
-  np. `void MyService.myMethod(String param1, Boolean param2)`
-- [ ] po wywołaniu metody logować informację o metodzie (w tym samym formacie co przed wywołaniem) wraz z informacją na
-  temat zwróconej wartości (wystarczy jej toString())
-
-### Wymagania techniczne
-
-- [ ] aspekt powinien zostać zaimplementowany z użyciem biblioteki AspectJ
-- [ ] rozwiązanie powinno spełniać zasady SOLID
-- [ ] testy jednostkowe rozwiązania nie są wymagane
-- [ ] testy integracyjne rozwiązania nie są wymagane
-- [ ] rozwiązanie powinno być odpowiednio zhermetyzowane (nie udostępniać funkcjonalności pozostałym pakietom programu)
-- [ ] kod powinien być odpowiednio udokumentowany za pomocą JavaDoc
